Writing findAverage test

+++ src/main/kotlin/findAverage.kt
@@ -1,5 +1,5 @@
-fun findAverage(input : List<Int>) : Float {
+fun findAverage(input : List<Int>) : Double {
     val sumTotal = input.reduce { count, number -> count + number }
     val size = input.size
-    return (sumTotal / size).toFloat()
+    return sumTotal/size.toDouble()
 }
